get some help reorganizing annoying files
